+++ .eslintrc.json
@@ -1,35 +1,15 @@
 {
-  "env": {
-    "browser": true,
-    "es2021": true
-  },
   "extends": [
-    "eslint:recommended",
-    "plugin:react/jsx-runtime",
-    "plugin:react/recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:react-hooks/recommended",
-    "plugin:jsx-a11y/recommended"
+    "next/core-web-vitals",
+    "next/typescript",
+    "plugin:jsx-a11y/recommended",
+    "plugin:react-hooks/recommended"
   ],
-  "overrides": [],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaVersion": "latest",
-    "sourceType": "module"
-  },
-  "plugins": ["react", "react-hooks", "@typescript-eslint", "jsx-a11y"],
+  "plugins": ["jsx-a11y", "react-hooks"],
   "rules": {
-    "react/prop-types": "off",
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "error",
-    "react/react-in-jsx-scope": "off",
-    "indent": [1, 2, { "SwitchCase": 1 }],
     "quotes": [1, "double", { "allowTemplateLiterals": true }],
     "semi": [1, "always"]
-  },
-  "settings": {
-    "react": {
-      "version": "detect"
-    }
   }
 }
