--- index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html class="dark" lang="en">
-  <head>
-    <link rel="shortcut icon" href="#" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,400;0,700;1,400;1,700&display=swap"
-      rel="stylesheet"
-    />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:wght@100"
-      rel="stylesheet"
-    />
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Armand's Web CV</title>
-  </head>
-  <body>
-    <div id="root" role="presentation"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
