+++ package.json
@@ -1,34 +1,28 @@
 {
   "name": "web-cv",
+  "version": "0.1.0",
   "private": true,
-  "version": "0.0.0",
-  "type": "module",
   "scripts": {
-    "dev": "vite",
-    "host": "vite --host",
-    "build": "tsc && vite build",
-    "preview": "vite preview",
-    "lint": "eslint --max-warnings 0 src",
-    "build-static": "renderToHtml.tsx"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "next": "14.2.8",
+    "react": "^18",
+    "react-dom": "^18"
   },
   "devDependencies": {
-    "@types/react": "18.3.5",
-    "@types/react-dom": "18.3.0",
-    "@typescript-eslint/eslint-plugin": "8.4.0",
-    "@typescript-eslint/parser": "8.4.0",
-    "@vitejs/plugin-react": "4.3.1",
-    "autoprefixer": "10.4.20",
-    "eslint": "8.57.0",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.8",
     "eslint-plugin-jsx-a11y": "6.10.0",
-    "eslint-plugin-react": "7.35.2",
     "eslint-plugin-react-hooks": "4.6.2",
-    "postcss": "8.4.45",
-    "tailwindcss": "3.4.10",
-    "typescript": "5.5.4",
-    "vite": "5.4.3"
+    "postcss": "^8",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5"
   }
 }
