--- src/ThemeContext.ts
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-type Theme = "dark" | "light";
-
-export const ThemeContext = createContext<Theme>("dark");
