+++ src/app/components/TimelineSegment.tsx
