+++ src/app/hooks/useOutsideClickHandler.ts
