+++ src/app/layout.tsx
@@ -1,0 +1,19 @@
+import type { Metadata } from "next";
+import "./index.css";
+import "./fonts.css";
+
+export const metadata: Metadata = {
+  title: "Armand's Web CV",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en" className="dark">
+      <body className={`antialiased`}>{children}</body>
+    </html>
+  );
+}
